ml/freshvision/requirements.txt
# FreshVision (EfficientNet) API
# Based on https://github.com/devdezzies/freshvision
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
